[#1496] Simplify legacy globals
Also adds removal instructions.
