+++ lib/onetime.rb
@@ -80,4 +80,3 @@ require_relative 'onetime/mail'
 require_relative 'onetime/alias'
 require_relative 'onetime/ready'
 require_relative 'onetime/boot'
-require_relative 'onetime/legacy'
